Custom Commands: move agent integration tests to separate file
These tests are repeatedly failing for unrelated changes, introducing
noise in other PRs. This PR isolates the tests in hope to make them be
more stable.
