+++ agent/src/__tests__/custom-commands/.cody/ignore
@@ -1,0 +1,2 @@
+# NOTE: For agent integration test only
+**/*Ignored.ts
