+++ agent/src/__tests__/custom-commands/src/animal.ts
@@ -1,0 +1,7 @@
+/* SELECTION_START */
+export interface Animal {
+    name: string
+    makeAnimalSound(): string
+    isMammal: boolean
+}
+/* SELECTION_END */
