+++ agent/src/__tests__/custom-commands/src/example2.ts
@@ -1,0 +1,3 @@
+export function example(): string {
+    return 'example'
+}
