+++ agent/src/__tests__/custom-commands/src/example3.ts
@@ -1,0 +1,3 @@
+export function example(): string {
+    return 'example'
+}
