+++ agent/src/__tests__/custom-commands/src/is_ignored.ts
@@ -1,0 +1,6 @@
+/* SELECTION_START */
+function isAlsoIgnoredByCody() {
+    console.log('This file is ignored by Cody via .cody/ignore.')
+    console.log('Used in testing for confirming the ignore rule is NOT case sensitive')
+}
+/* SELECTION_END */
