+++ agent/src/index.test.ts
@@ -19,12 +19,7 @@ import { TestClient, asTranscriptMessage } from './TestClient'
 import { TestWorkspace } from './TestWorkspace'
 import { decodeURIs } from './decodeURIs'
 import { explainPollyError } from './explainPollyError'
-import type {
-    CustomChatCommandResult,
-    CustomEditCommandResult,
-    EditTask,
-    Requests,
-} from './protocol-alias'
+import type { Requests } from './protocol-alias'
 import { trimEndOfLine } from './trimEndOfLine'
 
 const workspace = new TestWorkspace(path.join(__dirname, '__tests__', 'example-ts'))
@@ -876,177 +871,6 @@ describe('Agent', () => {
             )
 
             expect(client.textDocumentEditParams).toHaveLength(1)
-        }, 30_000)
-    })
-
-    describe('Custom Commands', () => {
-        // Skipped because this test an unstable prompt that keeps making the inline assertion fail.
-        it.skip('commands/custom, chat command, open tabs context', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            // Note: The test editor has all the files opened from previous tests as open tabs,
-            // so we will need to open a new file that has not been opened before,
-            // to make sure this context type is working.
-            const trickyLogicUri = workspace.file('src', 'trickyLogic.ts')
-            await client.openFile(trickyLogicUri)
-
-            const result = (await client.request('commands/custom', {
-                key: '/countTabs',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result?.chatResult as string)
-            expect(trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')).toMatchInlineSnapshot(
-                `
-              "Here are the file names you have shared with me so far:
-
-              1. \`src/TestLogger.ts\`
-              2. \`src/TestClass.ts\`
-              3. \`src/sum.ts\`
-              4. \`src/squirrel.ts\`
-              5. \`src/multiple-selections.ts\`
-              6. \`src/Heading.tsx\`
-              7. \`src/example.test.ts\`
-              8. \`src/ChatColumn.tsx\`
-              9. \`src/animal.ts\`
-              10. \`src/trickyLogic.ts\`"
-            `,
-                explainPollyError
-            )
-        }, 30_000)
-
-        it('commands/custom, chat command, adds argument', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-            const result = (await client.request('commands/custom', {
-                key: '/translate Python',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result?.chatResult as string)
-            expect(trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')).toMatchInlineSnapshot(
-                `
-              "Here's the translation of the provided TypeScript code into Python:
-
-              \`\`\`python
-              class Animal:
-                  def __init__(self, name: str, is_mammal: bool):
-                      self.name = name
-                      self.is_mammal = is_mammal
-
-                  def make_animal_sound(self) -> str:
-                      raise NotImplementedError("Subclasses must implement this method")
-              \`\`\`
-
-              In Python, we don't have interfaces like in TypeScript. Instead, we use classes and inheritance to achieve a similar functionality.
-
-              - The \`Animal\` class is defined as a base class.
-              - The \`__init__\` method is the constructor, where we initialize the \`name\` and \`is_mammal\` attributes.
-              - The \`make_animal_sound\` method is defined, but it raises a \`NotImplementedError\` to indicate that subclasses must override and implement this method.
-
-              Subclasses of \`Animal\` would then inherit from this base class and implement the \`make_animal_sound\` method with their specific logic.
-
-              For example:
-
-              \`\`\`python
-              class Dog(Animal):
-                  def make_animal_sound(self) -> str:
-                      return "Woof!"
-
-              class Cat(Animal):
-                  def make_animal_sound(self) -> str:
-                      return "Meow!"
-              \`\`\`
-
-              Here, the \`Dog\` and \`Cat\` classes inherit from \`Animal\` and provide their own implementation of the \`make_animal_sound\` method."
-            `,
-                explainPollyError
-            )
-        }, 30_000)
-
-        it('commands/custom, chat command, no context', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-            const result = (await client.request('commands/custom', {
-                key: '/none',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result.chatResult as string)
-            expect(trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')).toMatchInlineSnapshot(
-                `"no"`,
-                explainPollyError
-            )
-        }, 30_000)
-
-        // The context files are presented in an order in the CI that is different
-        // than the order shown in recordings when on Windows, causing it to fail.
-        it('commands/custom, chat command, current directory context', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-            const result = (await client.request('commands/custom', {
-                key: '/countDirFiles',
-            })) as CustomChatCommandResult
-            expect(result.type).toBe('chat')
-            const lastMessage = await client.firstNonEmptyTranscript(result.chatResult as string)
-            const reply = trimEndOfLine(lastMessage.messages.at(-1)?.text ?? '')
-            expect(reply).not.includes('.cody/ignore') // file that's not located in the src/directory
-            expect(reply).toMatchInlineSnapshot(`"9"`, explainPollyError)
-        }, 30_000)
-
-        it('commands/custom, edit command, insert mode', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(sumUri, { removeCursor: false })
-            const result = (await client.request('commands/custom', {
-                key: '/hello',
-            })) as CustomEditCommandResult
-            expect(result.type).toBe('edit')
-            await client.taskHasReachedAppliedPhase(result.editResult as EditTask)
-
-            const originalDocument = client.workspace.getDocument(sumUri)!
-            expect(trimEndOfLine(originalDocument.getText())).toMatchInlineSnapshot(
-                `
-              "// hello
-              export function sum(a: number, b: number): number {
-                  /* CURSOR */
-              }
-              "
-            `,
-                explainPollyError
-            )
-        }, 30_000)
-
-        it('commands/custom, edit command, edit mode', async () => {
-            await client.request('command/execute', {
-                command: 'cody.search.index-update',
-            })
-            await client.openFile(animalUri)
-
-            const result = (await client.request('commands/custom', {
-                key: '/newField',
-            })) as CustomEditCommandResult
-            expect(result.type).toBe('edit')
-            await client.taskHasReachedAppliedPhase(result.editResult as EditTask)
-
-            const originalDocument = client.workspace.getDocument(animalUri)!
-            expect(trimEndOfLine(originalDocument.getText())).toMatchInlineSnapshot(`
-              "export interface Animal {
-                  name: string
-                  makeAnimalSound(): string
-                  isMammal: boolean
-                  logName(): void {
-                      console.log(this.name);
-                  }
-              }
-
-              "
-            `)
         }, 30_000)
     })
